[util_lib] move cShake hash calculation into the common location
what happened

BRANCH=none
BUG=b:
TEST=manual
