+++ sw/host/provisioning/cp/BUILD
@@ -26,6 +26,5 @@ rust_binary(
         "@crate_index//:clap",
         "@crate_index//:humantime",
         "@crate_index//:log",
-        "@crate_index//:tiny-keccak",
     ],
 )
