+++ sw/host/tests/manuf/cp_provision_functest/BUILD
@@ -13,13 +13,13 @@ rust_binary(
         "//sw/host/opentitanlib",
         "//sw/host/provisioning/cp_lib",
         "//sw/host/provisioning/ujson_lib",
+        "//sw/host/provisioning/util_lib",
         "@crate_index//:anyhow",
         "@crate_index//:arrayvec",
         "@crate_index//:clap",
         "@crate_index//:humantime",
         "@crate_index//:log",
         "@crate_index//:rand",
-        "@crate_index//:tiny-keccak",
         "@crate_index//:zerocopy",
     ],
 )
